requirements.txt
docutils==0.14
Eel==0.10.4
flake8==3.7.8
flake8-polyfill==1.0.2
h5py==2.9.0
html5lib==0.9999999
imagesize==1.1.0
imutils==0.5.2
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.1.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lxml==4.3.2
matplotlib==3.0.3
mccabe==0.6.1
mkl-fft==1.0.0
mkl-random==1.0.1
mpi4py==2.0.0
mpld3==0.3
numpy==1.16.2
numpydoc==0.8.0
opencv-python==4.0.0.21
packaging==19.0
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
pickleshare==0.7.5
Pillow==5.4.1
playsound==1.2.2
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.4.7
py==1.8.0
pygame==1.9.6
Pygments==2.3.1
PyInstaller==3.5
pylint==2.2.2
PyQt5-sip==4.19.14
pytest==2.9.1
python-dateutil==2.8.0
radon==3.0.3
requests==2.21.0
rope==0.12.0
scikit-learn==0.21.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.9.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
